Add files via upload
Encounter represent a visit to the doctor’s office. The vital signs timeline is linked 
to the encounters. Each Encounter will have a single Vital Sign component.
The objective is to calculate the total number people in the community with 
blood pressure that is abnormal for any of the age groups. Create data for at 
least two communities and display abnormal cases for each community.